+++ packages/components/src/calendar/date-calendar/test/index.tsx
@@ -20,9 +20,14 @@ import {
 import { ar } from 'date-fns/locale';
 import { useState, default as React } from 'react';
 import '@testing-library/jest-dom';
-import { DateCalendar, TZDate } from '..';
-import { getDateButton, getDateCell, queryDateCell, monthNameFormatter } from '../utils/test-utils';
-import type { DateCalendarProps } from '../types';
+import { DateCalendar, TZDate } from '../../';
+import {
+	getDateButton,
+	getDateCell,
+	queryDateCell,
+	monthNameFormatter,
+} from '../../utils/test-utils';
+import type { DateCalendarProps } from '../../types';
 
 const UncontrolledDateCalendar = (
 	props: DateCalendarProps & {
